+++ src/api/auth/signup/signupAPI.ts
@@ -1,12 +1,9 @@
-import axiosInstance from '@/api/axiosInstance';
 import { SignupRequestParams, SignupResponseDto } from '@/types/signupType';
+import axios from 'axios';
 
 export const signupAPI = async (
   userData: SignupRequestParams
 ): Promise<SignupResponseDto> => {
-  const response = await axiosInstance.post<SignupResponseDto>(
-    '/auth/register',
-    userData
-  );
+  const response = await axios.post<SignupResponseDto>('/api/signup', userData);
   return response.data;
 };
